feat(perf): Landing change frontend tabs
To avoid confusion due to specific performance terms (pageload etc.) we're going to split the frontend tabs into both 'Web Vitals' and a general 'Frontend' tab. This way it's more clear that one view is specific to improve web vitals, which is only a portion of frontend performance. The other tab can handle overall frontend performance, including 'interactions' if we add them in the future.