feat(perf): Remove tab stickiness for landing
The 'stickiness' of specific views is confusing when switching projects.
This change will cause switching projects to automatically go to the
preferred performance view instead of retaining the previous one. We
originally were going to keep 'All Transactions' sticky between project
switches, but it would be an edge case that would likely be confusing
for users.