feat(sentryapps): Fallback to the default SentryAppAvatar if needed (#30461)
This PR will address a niche, rare bug that only happens if the stored image for an avatar is inaccessible. The default behavior for BaseAvatar is to fallback to the LetterAvatars. Instead, BaseAvatar (which already has considerations for checking if there should be a backup) will accept a 'backupAvatar' prop, and use that instead.