fix(dashboard-tests): Fix timeout exceeded failures (#30502)
The 'can enter edit mode for widgets' test was failing
with timeouts on CI.

So it seemed like the async openModal calls on the other
tests in this module had lingering side effects which were
causing the timeouts. Removed the asyncModal stuff in
favour of spying on the openModal function to ensure the
modal openend. We lose some coverage which ensured
that the PUT/POST mocks were called with the right
payload. For now, enhanced the modal tests to make
sure the "onSubmit" on these modals are called with
the right info.