fix(auth): Create new org membership in automatic IdP migration (#30356)
Previously, users attaching an AuthIdentity to an existing account via
idpmigration.py could do so only if they already were a member of the
provider's organization and would hit an OrganizationMember.DoesNotExist
exception otherwise. Modify the flow to dispatch correctly to
handle_attach_identity, creating new entities of both AuthIdentity and
OrganizationMember when needed.

Refactor AuthIdentityHandler to take the identity mapping on init and
store the User with a matching email address as an instance variable.
Introduce a property to distinguish that user from the authenticated
user attached to the request, if any.