feat(ratelimit): Ratelimit keys for authtokens & apikeys (#30423)
Objective:
We want to able to observe ratelimits for requests made using authtokens (sentryapps and user) and apikeys
Currently we handle AuthToken/ApiKey authentication during an Endpoint's dispatch(), which is after our ratelimit middleware. This PR brings the authentication logic up to the auth middleware.

However, we will not reject requests here. This is outside the scope of this PR. Rejection will still occur during the dispatch(). The auth middleware will swallow all the errors and let the request continue on.