debug views use notification logic (#30619)
This PR changes the debug emails for organization join and invite requests to use the actual notification to generate the email views. The purpose of this is so we have a single source of truth for the logic between debug views and the notification. Otherwise, we'd be in danger of the debug views not working after someone updates a notification but forgets to update those views. It also makes it easier to test changes notifications without having to actually trigger the email.