fix(post_process): Prevent auto assignments from ping-ponging between teams. (#30609)
In rare cases auto assignments can ping-pong between teams. This happens when multiple events with
different stack traces arrive within a minute window. Since we cache the result of the check for an
assignee existing for 60 seconds when one doesn't exist, the problem is that every subsequent event
sees a stale value and attempts to perform assignment again.

This could happen without the caching though - if two events were being processed simultaneously
then we would likely see this same problem.

To fix this, I have added `create_only` to the `assign` method. This causes assignment to halt if we
find that the group already has an assignee, and since it's protected by constraints in the database
we can rely on this working correctly.