include recipient type in referrer (#30643)
This PR adds the recipient type (team or user) to the referrer for links to Sentry coming from certain types of notifications. This is so we can track if the notification went to a team or a user when they load up the UI in Sentry. Note that this PR does not add the recipient type for all notification links yet but that will happen in a future PR.