adds secondary analytics event optionally (#30645)
This PR is part of a greater effort to increase standardization of notifications by adding support for two different analytics events for sending a notification. Now we will always record the integrations.{provider.name}.notification_sent event for every notification in addition to an optional one based on analytics_event. The purpose of this is to make it easier to do different types of analysis (e.g: looking at a specific notification funnel vs tracking how many times we send any type of Slack notification).