fix(ui): Span tree connector design (#30498)
* fix(ui): Span tree connector color

The current color for span tree connectors is gray200. This is too light. Instead, we should use gray300 at 0.4 opacity.

* ref(ui): Increase line width for span tree connectors

The current line color is gray200. This is too light when the line width is 1px. Increasing the width to 2px makes the line more prominent.

* fix(ui): spanGroupBar offset

There is a slight offset on spanGroupBar (see PR #30498). This commit fixes that.

* fix(ui): TreeConnector offset by 1px

* fix(ui): Fix TransactionBar connector offset

* fix(ui): Fix TransactionBar connector offset