ref(notifications): consolidate notification logic (#30575)
This PR moves some logic and methods from the OrganizationRequestNotification class to its parent BaseNotification class. This is part of an effort to ensure greater standardization of notifications. Functionally, the only change is modifying the referrer for links to in the format of referrer_base + provider_type + recipient_type. This will improve our analytics so we know which provider and which kind of recipient are going to links in Sentry. It also changes the format slightly in some cases (AlertRuleSlackTeam becomes alert-rule-slack-team). However, this change will not impact how we look at the referrer in analytics.