fix(issues): Fix issue where `get_event_by_id` can fail to return details of an existing event
When we call `get_event_by_id` without a `group_id`, we make a call to snuba to fetch the event
details so that we can figure out the `group_id`. If this method is called after an event is
created, but before it exists in snuba, then we cache the empty result for 7 minutes. This then
results in these calls returning no event until the cache expires.

To work around this, I add a useless condition that compares against a random timestamp from before
2002. This means that we will effectively never hit the cache for this query. This adds a small
amount of load to clickhouse, but shouldn't cause any problems. The ideal solution here would be to
just not negative cache for this particular referrer, but that could be a longer term solution.