feat(doc integrations): Add GET endpoint for all default IntegrationFeatures (#30769)
See API-2333

This PR adds an endpoint to view all the default IntegrationFeatures. Since it lives in Python and not the DB, My work around to "serialize" it was to create an IntegrationFeature in memory, and serialize that one without any FK fields. I could have just manually assembled the dictionary in the endpoint, or created a new serializer just for the Features enum, but I thought this was the best compromise.