feat(integrations): Fetch integrations pending deletion and block reinstalls until deleted (#30686)
Objective:
Currently, when we uninstall an integration, in the UI it looks like its automatic. But we are actually scheduling the OrganizationIntegration for deletion and then cascading it. This task happens every 15min.

But customers are immediately re-installing the integration. This causes issues with associating repositories and our code/user/team mappings.

This PR returns the integrations pending deletion and will block new configurations for an integration if an existing one is pending deletion.