feat(insights): Allow a list of statuses to be passed to `TeamIssueBreakdownEndpoint` (WOR-1511) (#30801)
This allows a list of statuses to be passed to the issue breakdown endpoint. If no list of statuses
is passed, the old behaviour will be used, where we return the data for each bucket like:
`{"total": 0, "reviewed": 0}`

If a list of statuses is passed, then reviewed will be replaced by the specifically requested
statuses instead. So if `["new", "unresolved", "regressed"]` was passed, each bucket would look like
`{"total": 0, "new": 0, "unresolved": 0, "regressed": 0}`