chore(sentry-apps): Add avatars to the SentryAppsStats endpoint (#30799)
See API-2332

Jokes on me for thinking this would only be 2 PRs. In order to add the avatars in _admin tools, we'll need access to the avatars from this endpoint. It is already restricted to SuperUsers only so there isn't functionality change besides extra fields and we're not gonna be leaking anything. Tests have also been updated to reflect the change.