fix(insights): Avoid double counting for a group when determining unresolved issues.
There are duplicate status rows in the `GroupHistory` table. This can happen with a group being
resolved multiple times in a row, or really another other status other than `NEW`.

To make these stats correct we want to only include a `GroupHistory` row in the count if it's not a
duplicate. To determine a duplicate we:
 - Query for the most recent relevant `GroupHistory` row related to the same group that is older
   than the current row. Relevant means that it's either an OPEN or CLOSED status. So we don't care
   about ASSIGNED/UNASSIGNED, but we do care about UNRESOLVED, REGRESSED, RESOLVED, etc.
 - We then look at the status of this row and determine whether it's a CLOSED or OPEN status. We do
   the same for the current row, and if both are either CLOSED or OPEN then this row is a duplicate,
   so we discard it.

This will make this report less efficient, and the way that we have to do this in Django means that
the subquery is repeated multiple times. There are more efficient ways to write this if we break out
the raw sql, but for now this will work and we can see if we end up having performance problems
here.