feat(doc integrations): Only allow publishing when there is an avatar (#30806)
While working on a ticket to update the avatar, I noticed that there was no validation in place to prevent publishing without an avatar. This check just throws validation errors when trying to publish if there is no record of an avatar. Updates to the DocIntegration that don't involve publishing still work as expected.