fix(projectconfig): Fix bug where projectkey deletion would not trigger
In order to improve performance for projects containing a ton of project
keys, refactor code such that project-id-based cache keys are no longer
written, and such that a single key is written when a pk is updated
rather than all keys related to the project. This breaks a bunch of unit
tests, which in turn uncover two logic errors in how deletion or update
of project keys is handled.

The customer-visible problem is one of two:

* Have so many keys per project and don't observe updates to eg key
  quota at all, as tasks time out.

* Disable or delete project key, which is not observed in time in Relay.

Disabling of pk: In master we intentionally skip projectkeys that are
not active. We do this to make the projectconfig smaller, and that made
sense back when projectconfig cache keys were projectid based. This is
no longer the case since the refactor.

Deleting of pk: When a pk is deleted, it can't be found in the update
task, and the task crashes. Handle this special case (see code comments
about limitations).