feat(doc integrations): Use docIntegrations from the backend with feature flag (#30903)
See API-2324

This PR will adapt the frontend to use the DocIntegrations returned from the backend rather than the hardcoded list in the code right now (if flagged in). Without the flag, the it falls back to the regular list and renders exactly as expected.

Additionally, this PR also makes docIntegrations and sentryApps appear properly in search (with their rendered avatars).