feat(ratelimits): Add headers with rate limit details
The headers allow API users to know where they are in terms of their rate limits. It'll be returned for every API that can be rate limited except when there is an internal exception.
At the time of this commit, rate limits are not enforced except for some specific endpoints.