ref(pageFilters): Do not navigate sidebar w/ statsPeriod
This removes the `navigateWithGlobalSelection` functionality from the
sidebar.

Previously the behavior was that the `statsPeriod` would carry over if
changed between pages, as well as project and environment, however these
are stored in local storage currently, so these will all automatically
carry over anyway.

The start, end, and utc parameters when using absolute dates *were not
currently carried over* due to a bug in the default parameters of
the PageFiltersContainer, causing these parameters to never carry over.

The sum of this change will simply be that `statsPeriod` no longer
carry over.