feat(dashboard): Add a JSONField to store widget details
We have a few columns on our DashboardWidget model
that are *descriptors* of sorts for what a dashboard
widget might look like, for example: `display_type`,
`widget_type`, `interval`...

We anticipate adding more descriptors in the future,
like widget sizes, maybe widget coordinates, table
widget column widths, top events limit etc. and it
would be nice to not have to create a field and
migration for each of these on this model.
Instead, we can put these in a JSON field.