feat: Add pytest-xdist to dev requirements
While not particularly required, this introduces a corrected `LiveServerTestCase`. This dynamically adjusts `system.url-prefix` to be correct.

This removes the use of the initialize fixture on symbolicator tests as it is incompatible with `pytest-xdist` (and also quite confusing). As part of that change it refactors the symbolicator test suite to be a bit more manageable.