feat(ci): Balance frontend tests (#30949)
This adds a jest reporter to record the test durations of our frontend tests. It also changes our runner to look for a results JSON and attempt to balance tests between our CI runners so that each runner finishes around the same time. Previously, the tests were split by the # of test files (so each runner tested the same # of test *files*).

Ideally we would add some automation to update the balance JSON file. But even without automating it, this should suffice for awhile (until we add new tests that are quite slow).

Some automation ideas:

* Scheduled workflow that runs the frontend tests and generates the results json
* Generate the json while running tests, upload as GHA artifacts, and for each test run, pull from GHA artifacts and merge/generate the json (since test runs are sharded, they will have separate result json)