fix(insights): Exclude ignored issues from initial unresolved counts in all unresolved issues endpoint (#31053)
Since we've only been writing to `GroupHistory` in the last few months it doesn't have a full
history going back to when all groups were created.

Currently, if a group was ignored before we started writing `GroupHistory` rows then it is included
in the total counts, which throws off our daily numbers. This pr tries to detect these rows, by
filtering to groups that are ignored and have no relevant rows in `GroupHistory`.

Unfortunately we can't use `GroupSnooze` for this, since it has no `date_added` field.