fix(dashboards): remove last field from topn fields when opening in discover (#31259)
Removes the last field from Top N Widget fields when clicking Open In Discover. This is because the last field for Top N widgets is used as the Y Axis, so we remove it before passing to discover, otherwise discover would render the Y Axis as an extra column in the table display.