fix(sessions): Order results by timestamp and log error if snuba limit exceeded (#31214)
As described in https://getsentry.atlassian.net/browse/ISSUE-1372, gaps
occur in sessions_v2 time series when the number of releases is large.
This seems to be caused by the fact that snuba applies a default limit
of 1000. The sessions API queries these series without an orderBy
constraint, so a random subset of entries default to zero.

This PR logs an error if this limit is actually reached. Furthermore, we
add an order by clause to the snuba query, such that at least the most
recent part of the time series is complete.