feat(dashboards): Force Issue field on Issue widgets and remove redundant permalink field (#31173)
Forces the Issue field to always be displayed on Issue widgets. This is because Issue table results are always grouped by the Issue # regardless of fields selected so we should always display the Issue #.
If there is only one Issue column, it is disabled/unchangeable and undeletable.

Permalink is no longer selectable in the field dropdown because it is functionally the same as the Issue column, therefore redundant.