feat(ownership): Add feature to allow some orgs to have a higher max project ownership limit (#31231)
We have this limit in place to protect our infrastructure against problems caused by a large number
of ownership rules.

In some cases, we might want to relax this for specific orgs. This adds in a feature flag to allow
for a higher limit.

Longer term we should figure out how to make processing large number of rules more efficient. We can
probably do prefix matching for a lot of these rules, so that we don't have to scan every rule in
the tree for every path in the stack trace.