fix(api): Fix OrganizationTagKeyValuesEndpoint to support passing environment
We pass in environment name to `get_tag_value_paginator_for_projects`, but we actually require an id
here. When passing an environment name we actually error here when we attempt to translate the
environment filter param.

This fix just changes the code to pass environment ids rather than names.