fix(consumer-lag): Remove header optimization for some cases
When the header `transaction-forwarder` is present in the kafka headers,
snuba errors and transaction consumers apply an optimization where
they ignore messages which are not destined for them. But the way
they ignore it cause offsets not to be committed often if there are
only 1 type of messages on the shared topic. Removing the header if the
option "eventstream:kafka-headers" is not set. This would resolve
https://github.com/getsentry/self-hosted/issues/1204