fixed /stats_v2/ endpoint returns zero for a lot of time slices (#31296)
problem: https://getsentry.atlassian.net/browse/ER-913

The reason why this was happening was in make_timeseries, how the rows
was sorted was changed. Tbjs caused the row[ts_col] == ts to not match
in most cases. In turn this set row to None which resulted in a 0 in the
series field for the date that it was compared to.

The fix is to revert it back to how the rows were being sorted before.