fix(ui): Handle project selection when logger clicked
Fix issue where when you click the logger from the Issue Stream, it would go to the Custom Search tab but remove the project from the global header. If you don't have the multiple project feature, it would display an error. This fix should persist the selected project.

FIXES ISSUE-1381