ref(discover): Dataset-specific config to specify fields and filters (#31272)
Introduce a dataset-specific configuration that generates
the available fields, functions and filters for the dataset
so that the logic doesn't live inside QueryBuilder. Furthermore,
remove the multiple levels of inheritance and move the
related functions inside QueryBuilder since a lot of
complexities are abstracted away by the dataset config.
Note, not removing the QueryFilter, QueryFields and QueryBase
class yet because they are inherited or called elsewhere.
Will be removed as a follow up.