ref(codeowners): Allow associations endpoint to be limited by provider (#31319)
See [API-2387](https://getsentry.atlassian.net/browse/API-2387)

This PR will introduce a query parameter option to the associations endpoint so that we limit the results based on provider. This will be useful in the team/user mappings pages since we don't want to show 'gitlab' user suggestions on the 'github' configuration page.

Tests have been added as well to ensure this feature works as expected.