feat(anomalies): Add anomalies tab backed by local storage
This will allow UI for the upcoming anomalies feature to work behind it's feature flag, and have response data supplied to it via local storage while the api is still being finalized. Temporarily does transformation of the api payload. Have some basic tests we can expand on in the future.