fix(insights): Improve all unresolved issues when we have 90 days of history data.
If we start at today, use the current count of unresolved issues and then move backwards through
history we'll get a more accurate count of unresolved issues, because we don't need to do estimates
from before we had history. I'll trial this on Sentry first and then remove the estimation logic
entirely.