ref(metrics): Align number of groups for sessions_v2 queries (#31389)
This PR prepares for #31392, where we want to limit the number of groups
for which the sessions V2 API queries time series. For the sessions
backend of release health, this is a straightforward change, but the
metrics implementation currently fires multiple requests with different
group by clauses at snuba.

This PR ensures that all metrics-based totals queries use the same group
by clause, hopefully simplifying the necessary limiting logic.