fix(dashboards): lazyload widgets into view when deleting upper widget (#31426)
When deleting a widget in a dashboard, the layout will shift to fill in the deleted widget gap. This can cause widgets that were previously out of view to come into view after layout shift. Our lazyload lib doesn't flag this event to check for lazyloading so we need to manually check by calling forceCheck(). Unfortunately we also have to use a setTimeout because RGL has an animation when changing layout (which doesn't seem to be configurable). RGL also doesn't have a handle for post animation layout changes so we don't have a more elegant way to schedule the forceCheck().