feat(perf): Improve Trace View when there are only errors and no transactions (#31401)
On the trace view, it is sometimes possible that a trace can contain only errors and no transactions. In those cases, an alert that says "The trace you are looking for only contains errors." is rendered. This PR improves the trace view in these cases, by rewording this message, and listing the error events in the trace with links to view them.

Fixes VIS-1385