fix(ratelimit): Store header info with request
Previously we stored the rate limit header information in a class property of the middleware assuming that each request had it's own instance of the middleware. Turns out that a lot of the requests share the same middleware so every request was thrashing the metadata property.

Instead, I moved the header information to a property of the request. This way the correct header is tied to its request