fix(perf): Style placeholder component to fit into the widget container (#31458)
* fix(performance): Set overflow to auto for performance widget container

Set container overflow to auto so that Placeholder component's vertical overflow does not clip containers bottom border.

Fixes VIS-1372

* ref(widgets): Style placeholder instead of the widget Container

Add transparent border to placeholder and style it to have the same border radius so its overflow is invisible for users. Remove overflow styling for the container to allow popup overflow.