feat(ci): Fix pytest annotations not working
There were two problems, the plugin detection for `pytest_rerunfailures`
was not working and pytest was swalloing the errors.

The other problem seems to be, for whatever reason, the GH annotation
log was also being swallowed up (it was being printed to stderr).
Printing it to stdout seems to work?