tests(codeowners): Add tests for integration user/team mappings (#31527)
See API-2404

This PR adds tests for the IntegrationExternalMappings (The table component) and IntegrationExternalMappingForm (the inline form and modal form) components. Both of these components are the root for IntegreationExternalUserMappings and IntegrationExternalTeamMappings, so I figured adding tests there would be a bit redundant. If you have suggestions for more tests or rewrites, please let me know!