fix(post_process): Fetch buffered `times_seen` values and add them to `Group.times_seen`
In `post_process_group` we process issue alert rules and also ignored groups. Both of these can have
conditions that read from the `times_seen` value on the `Group`.

The problem here is that updates to `times_seen` are buffered and only written every 45s or so. This
means that most of the time when a `Group` goes through `post_process_group` it has an out of date
`times_seen` value. For infrequently updated groups, this can just mean that the count is -1. But
for high volume groups this could mean that we're considerably below the count.

To improve this, we read the current value from buffers and add it to `times_seen`. There's a
potential race condition here where we fetch the `Group`, and before we fetch the value from buffers
it is cleared, and so we miss out on the update. This should be infrequent enough that it's not a
problem, and either way we will be considerably more accurate most of the time.

I did some checks to make sure that we don't call `group.save` as part of post process and I
couldn't see anywhere that happens, so this should be safe.