feat(ui): creates sessionStorage wrapper (#31612)
This PR makes a wrapper around sessionStorage that won't error out when we can't modify session storage the same way we handle localStorage. It also makes sure that if window is ever undefined, it won't blow up (not sure if that's possible with Sentry but it's definitely possible with server-side rendering).