feat: makes task to send notifications (#31620)
This PR adds a way to kick off notifications as a task through the function async_send_notification. This function will take arguments and serialize them to pure JSON and pass them to a task called _send_notification. It has specific logic to handle models by figuring out the table name and the primary key so we can re-query it inside the task, instantiate the notification class, and call send. The purpose of this is so in case we can asynchronously send emails/slack messages without making the user wait.