Move access logs into middleware (#31653)
### Problem

The API access log used to live in the base endpoint however if any middleware intercepts the request (for example the rate limit middleware), the endpoint code will never be hit. We still want those logs to be created so we move the logger into middleware

### Solution
Move the access logs into middleware as well. Put it early enough in the middleware onion that it will log any API request that we care about