fix(sessions): Limit number of groups in sessions_v2 API (#31392)
Previously, the /sessions/ api would return an unlimited number of data
points, depending on the number of different groups existing in the
dataset (e.g. for 3 releases and 2 environments, group by release and
environment would return 6 groups).

Because time series are also grouped by timestamp, the number of data
points would easily exceed the default snuba limit of 1000 rows. In this
case, the api would simply return a zero or null value for any data
point not present in the snuba result (see
https://getsentry.atlassian.net/browse/ISSUE-1372).

This PR:

- Increases the limit on the snuba query to 5000.
- Limits the number of groups returned for each time series such that
  only the top-N groups are returned.
- N is determined by the number of intervals selected through
  statsPeriod and interval
- Groups are ranked in descending order by the value of the first field
  requested, e.g. -sum(session).