ref(access): Convert set attributes to derived properties of Access
Convert Access to an abstract class, changing teams and projects to
dervied properties, and the "has access" methods to abstract.

Introduce OrganizationMemberAccess to represent the former base case
where a user has access only to a subset of their organization's teams
and projects. Push has_global_access down to here.

Use OrganizationGlobalAccess in more places, to represent where
superusers and Sentry apps have access to all an organization's teams
and projects. Enumerate those sets of teams and projects as properties.

Tweak the class structure as needed.