fix(metric_alerts): Make sure critical triggers resolve properly when no action is set on a warning trigger
Problem:
If we have an alert set up like:
 - Warning: 50. Action: None
 - Critical: 100. Action: Slack

Then if we go from critical -> warning state the slack resolve action will fail to fire.

Cause:
The reason this happens is related to a previous fix. For an alert like
 - Warning: 50. Action: Slack
 - Critical: 100. Action: Slack

When going from critical -> warning the critical action would be marked as resolved. This would cause a slack notification with `Resolved` to be sent to the channel. This is misleading, because the alert is still active, just in the warning state. What we want here is to fire a warning notification instead.

The initial fix for this was that when we resolved a critical trigger, we’d check and see whether there is an active warning trigger. If so, we’d send a warning trigger fire to our actions, rather than a critical trigger resolve. This works ok for many cases, but fails when the actions on the warning trigger are different to those on the critical trigger.

Fix:
Substituting the warning trigger for the critical trigger causes us subtle bugs. So, instead of this, when triggering fires/resolves on our action handlers we will also pass along the incident state change that the trigger/resolve caused the incident to go into.

So if a critical trigger resolves, we check what state it would have put the incident in. If there’s a warning trigger, then the state is warning. If no warning trigger, the state is closed. This state is then used to appropriately generate the messages that we send to users via our various actions.

So now, If we have an alert set up like:
 - Warning: 50. Action: None
 - Critical: 100. Action: Slack

If this goes from
 - critical -> warning OR critical -> resolved we will send `IncidentStatus.WARNING` to any actions related to the critical trigger.
 - warning -> resolved We do nothing since there are no actions on the warning trigger

If we have an alert set up like:
 - Warning: 50. Action: Slack
 - Critical: 100. Action: Slack

If this goes from:
 - critical -> warning: critical trigger, `IncidentStatus.Warning`
 - warning -> resolved: warning trigger, `IncidentStatus.Closed`
 - critical -> resolved: Since we de-dupe triggers to avoid spamming the user, we will select the warning trigger here, and send `IncidentStatus.closed`

If we have an alert set up like:
 - Warning: 50. Action: Slack
 - Critical: 100. Action: Pagerduty

If this goes from:
 - critical -> warning: critical trigger, `IncidentStatus.Warning` sent to Pagerduty. Nothing sent to Slack
 - warning -> resolved: warning trigger, `IncidentStatus.Closed` sent to Slack. Nothing sent to Pagerduty
 - critical -> resolved: Critical trigger, `IncidentStatus.Warning` sent to Pagerduty. Warning trigger, `IncidentStatus.Closed` sent to Slack. We don’t de-dupe here since the actions are different.