fix(release_health): Split up featureflag [INGEST-784]
We have recently enabled 100% of metrics extraction for release health,
and so we want to start querying that data to know how much load that
puts on ClickHouse.

While we do not have a full/proper dataset for all organizations yet
(which affects query performance), we can still start sending queries
unconditionally and filter out reports by another featureflag.

After this PR is merged, some changes in Flagr are necessary to continue
receiving reports:

1. organizations:release-heath-check-metrics needs to be renamed (or duplicated) to the new featureflag organizations:release-health-check-metrics-report
2. the old featureflag needs to be enabled for _at least_ the organizations we want to compare results of, but we probably just want to enable it for all.