fix(metrics-api): Only paginate with orderBy (#31918)
We disallow pagination parameters per_page and cursor except when there
is also an orderBy. But the default per_page of 50 was still
accidentally applied as a limit in any other case, leading to incomplete
time series filled with zeros.

With this PR, the default limit provided by the paginator is now ignored
unless there is an orderBy.