fix(spans): Fix example spans query to return only events containing both span.op and span.group (#31855)
This PR adds an additional condition to the span example query checking that examples have the same span.op and span.group as the passed in span. It's possible that an event has the span.op but with a different span.group; thus, returning an event that isn't an example for a given span. This is a backend fix for #31552

Fixes VIS-1411

* fix(spans): Fix example spans query to return only events containing span.op and span.group

* install js dev

* chore(spans): Use consistent alias

* fixes

* test(spans): Test queries

* test(spans): Fix tests

* remove mocks

* add a comment

* fix tests

* fix assertion

* update response shape