fix(deletions) Fix CodeOwners breaking deletions
There are a few organization deletions stuck on the protected relation
with ProjectCodeOwners. When removing a repository we must remove the
RepositoryProjectPathConfig which means we also need to remove the
ProjectCodeOwner records. If we want to keep code owners configuration
around we need to make the foreign key nullable.

Fixes SENTRY-S4Y