feat(perf): Add anomalies backend and summary graphs (#32050)
* feat(perf): Add anomalies backend and summary graphs

This hooks up the actual anomalies backend and adds anomalies to the transaction summary tpm graph.

This is an pre-internal PR for the anomalies ui, still behind a feature flag, to get this onto prod for others to test before internal.
Still needs some extra work on tests etc. but this should be enough to play with for now.