fix(discover): Add more robust check for isMultiSeries
We had an issue where a topN (multi series stat) response
had `data` as the series name for one of the series. This
resulted in the isMultiSeriesStat check failing because
data.data === undefined evaluated to `false`. Single series
also returns start and end dates, so adding that to the check
to make the single series check more robust. Note that start
and end are optional props on EventsStat type, so this is not
the optimal solution. Long term solution is to update the API
response schema from the events-stats endpoint.