feat(issue_alert_status): Add endpoint for fetchcing group history for a issue alert rule
This adds `ProjectRuleGroupHistoryIndexEndpoint`, which fetches the history of groups that have
fired for an alert, filtered to a specific time range.

Also made attempts to properly document this endpoint. I've not marked it as public yet since it
might change as we develop the feature. Most of the documentation work was just documenting the
`GroupSerializer`.