feat(issue_alert_status): Add method to fetch issue alert group break down in a time range. (#32099)
This adds `fetch_rule_groups_paginated` to the `RuleHistoryBackend` classes. This returns a
`CursorResult` which contains a list of `RuleGroupHistory` objects as the result. These objects
encapsulate a `Group` and any extra context about them, which at the moment is just the number of
times the group triggered the rule.