fix(alertWizard): Chart labels to show correct date/time
statsPeriod was previously defaulting to 1d instead of 14d even though the Display selector was showing Last 14 days. Because it was defaulting to 1d, the date/time chart labels would show as one time across the axis on page load. This should now show the correct date/times and default to 14d on page load.

FIXES WOR-1649