feat(issue_alert_status): Add endpoint for fetching group history for an issue alert rule (#32143)
This adds `ProjectRuleGroupHistoryIndexEndpoint`, which fetches the history of groups that have
fired for an alert, filtered to a specific time range.

Also made attempts to properly document this endpoint. I've not marked it as public yet since it
might change as we develop the feature. Most of the documentation work was just documenting the
`GroupSerializer`.