fix(profiling): Set start/end time range on scatter chart
By default, echarts uses the min/max x values from the series to determine the
bounds of the chart. Because the scatter plot only contains data points, this
means the min/max timestamps will be used. This change enforces the min/max
value for the chart based on the page filters.