fix(perf): Fix sticky transaction duration in search
Modifying the search bar wasn't pushing context of an empty input since no query in the url passed the same checks we already did for default query. This adds an additional query param once a user has modified the search (so it can be linked without transaction duration default). This param is checked for before providing the transaction duration defaults. In the long term we will just remove the transaction defaults.