fix(metrics): Zero-fill response when there's no data [INGEST-941] (#32157)
When there isn't any metrics data, the `groups` of the response is empty.
However, the absence of data must be represented with an appropriate value.
For example, `count_unique(user)` must return `0` when there aren't any
users, instead of returning no data. The value representing the absence of
data is `0` for sums and counts, and `None` for everything else
(such as `p50`).