feat(dev): Do not write CPATH and LDFLAGS in .zshrc
First of all, updating it to correct values in the future is complicated.

Second, it had hard coded versions of the libraries. This means that future brew installations
could make it useless.

Exporting it as variables within the `make install-py-dev` target seems sufficient.

The negative of this approach means that `pip install` of these packages without using the make target would fail.